third_party/WebKit/LayoutTests/fast/dom/shadow/contains-with-shadow-dom-expected.txt

Tests for node's contains().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.contains(getNodeInShadowTreeStack("shadowHostA/")) is false
PASS getNodeInShadowTreeStack("shadowHostA/").contains(document) is false
PASS document.contains(getNodeInShadowTreeStack("shadowHostA/divA")) is false
PASS getNodeInShadowTreeStack("shadowHostA/divA").contains(document) is false
PASS getNodeInShadowTreeStack("shadowHostA/").contains(getNodeInShadowTreeStack("shadowHostA/divA")) is true
PASS getNodeInShadowTreeStack("shadowHostA/divA").contains(getNodeInShadowTreeStack("shadowHostA/")) is false
PASS getNodeInShadowTreeStack("shadowHostA/").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostB/")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostB/").contains(getNodeInShadowTreeStack("shadowHostA/")) is false
PASS getNodeInShadowTreeStack("shadowHostA/divA").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostB/divB")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostB/divB").contains(getNodeInShadowTreeStack("shadowHostA/divA")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostB/").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostC/")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostC/").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostB/")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostB/divB").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostC/divC")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostC/divC").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostB/divB")) is false
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostB").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostB")) is true
PASS getNodeInShadowTreeStack("shadowHostA/shadowHostB/").contains(getNodeInShadowTreeStack("shadowHostA/shadowHostB/")) is true
PASS successfullyParsed is true

TEST COMPLETE
